Fix rebasing conflicts
