+++ src/actions/articleActions.js
@@ -102,7 +102,7 @@ export const getSingleArticle = slug => (dispatch) => {
       if (response.success) {
         dispatch(getSpecificArticle(response.content));
       } else {
-        toastr.error(response.error.message);
+        dispatch(getSpecificArticle(response.error.message));
       }
     });
 };
@@ -111,8 +111,40 @@ export const createArticle = articleDetails => (dispatch) => {
   ArticleAPI.createArticle(articleDetails).then((response) => {
     if (response.success) {
       dispatch(createArticleSuccess(response.article));
+      dispatch(getSpecificArticle(response.article.slug));
     } else {
       dispatch(createArticleFailure(response.error));
+    }
+  });
+};
+
+export const deleteSpecificArticle = payload => ({
+  type: types.DELETE_ARTICLE_SUCCESS,
+  payload
+});
+
+export const editSpecificArticle = payload => ({
+  type: types.EDIT_ARTICLE_SUCCESS,
+  payload
+});
+
+export const deleteArticle = slug => (dispatch) => {
+  ArticleAPI.deleteArticle(slug)
+    .then((response) => {
+      if (response.success) {
+        dispatch(deleteSpecificArticle(response.article));
+      } else {
+        dispatch(deleteSpecificArticle(response.error.message));
+      }
+    });
+};
+
+export const editArticle = (slug, articleDetails) => (dispatch) => {
+  ArticleAPI.editArticle(slug, articleDetails).then((response) => {
+    if (response.success) {
+      dispatch(editSpecificArticle(response.article));
+    } else {
+      dispatch(editSpecificArticle(response.error));
     }
   });
 };
